feature: new version was publish
