+++ BoxAgileDevCore/Infrastructure/Extensions/HttpRequestMessageExtension.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using System.Net.Http;
+
+namespace BoxAgileDevCore.Infrastructure.Extensions
+{
+  public static class HttpRequestMessageExtension
+  {
+    public static HttpRequestMessage SetHeaders( this HttpRequestMessage requestMessage, IDictionary<string, string> headers )
+    {
+      foreach ( KeyValuePair<string, string> header in headers )
+      {
+        requestMessage.Headers.TryAddWithoutValidation( header.Key, header.Value );
+      }
+
+      return requestMessage;
+    }
+  }
+
+}
